Major rework and reorganization of example app.
Main fix is the excessive memory use caused by co-routine prompting and
a location where we should have been swapping values differently in
Action; also some reorganization of the code to a more useful
arrangement. There's still some more of that to do that would be useful
for reducing mutli-arch-related noise.