Decouple dev shell from crate2nix stuff
This allows one to enter the shell even if the `Cargo.toml` or
`crate-hashes.json` is invalid or out of sync.

The cost if that we now pin the SDK in a thunk *and* the lock file, but
it has been requested that we put the script logic in `cargo-ledger`
instead, which would mean this duplication goes away.